Generate DBus interfaces for client libraries
The generated dbus-proxies.h and dbus-proxy-mocks.h are now created by
by the daemon that exposses the interface, instead of letting the
daemon using the interface generate them. This creates then a client
library that other daemons can use.

This patch installs these libraries for login_manager, power_manager,
debugd, apmanager and shill, and makes all their callers use these
interfaces instead of using the xml directly.

This patch moves the code dependency on the .xml files to an ebuild
dependency.

This patch only contains the apmanager changes. The original CL is here:
https://chromium-review.googlesource.com/296026

BUG=23619937
TEST=./build_packages

Change-Id: I80855787fb22438bcc2e36eede987e8cc0d083e4