Build apmanager from the new directory path
The apmanager ebuild recreates the old directory structure placing
apmanager as a subdirectory of platform2 in the build environment.
This structure doesn't match the current location of the code and
breaks incremental builds creating symlinks to apmanager in
platform2/ and even in the new apmanager directory.

This patch modifies the .gyp file to reference platform2 sources via
the platform2_root gyp variable and builds apmanager from the new
path. This fixes incremental builds.

For this change to take effect in Chrome OS, the ebuild file is updated
in a separated CL.

Bug: 23619937
TEST=emerge-$BOARD apmanager; cros_workon_make apmanager

Change-Id: I44d5afd2b47ff6508e98b583a25fd92126484d90