Rename shill_proxy to shill_manager
The end goal is to provide an abstration interface for shill proxy,
to allow different implementations for the proxy.

One immediate beneifit is that we can use a stub implementation
to temporarily remove the dependency for libshill-client while
compiling apmanager on Android. In the future, we might provide an
implementation based on Binder for Android.

Bug: 24164800
TEST=emerge-$BOARD apmanager

Change-Id: I4b241aee97252704862eb77419cdce4a858e19a1