Do not run apmanager in minijail for now
When starting apmanager through Android's init system, the daemon
will be started under "system" user, and that's currently targeted
user for running apmanager. So there is no need to use minijail to
drop the user privileges to "system" user" again.

So skip this for now until we create a new user/group to run apmanager
under.

Bug: 22828115
TEST=Manual test on dragonboard

Change-Id: Ib37553126f9a9a4024a16f68d91d240ca855f55c