Do not use singleton for mock objects
Android complains about memory leakage when using singleton
for mock objects, and it is not really necessary to use
singleton for mock objects either.  So remove the singleton
from those objects.

While there, update the binary path for "sleep" on Android,
which is used for starting a dummy process in the test.

Bug: 25186794
TEST=Run apmanager_test on Brillo board
TEST=Run unittests for Chrome OS

Change-Id: Ifb5a5c6876aa67a94411d823109c14a209a46d86