brillo: start AP service after AP interface is enumerated
For Brillo devices, it is required to use the AP mode interface
that's setup using the HAL API. So explicitly setup an AP mode
interface using the HAL API, and wait for the interface to
enumerated before starting the AP service.

This is achieved by handling the Start request asynchronously.
Also explicitly update the AP Config to use the newly created
AP interface, to avoid the Config from auto selecting a wrong
interface in the case when multiple AP-capable interfaces are
presented.

Bug: 25113165
TEST=Device setup works on both Brillo boards

Change-Id: Icdee32e8bab9cf41a467058aef5ea081562a3442