Define and implement an AdaptorInterface for Config
This interface will be used by Config to communicate with
the adaptor that exposes its functionality over IPC.

The D-Bus implementation of this interface is included, and
integrated to the D-Bus control interface.

An upcoming CL will integrate this AdaptorInterface to the
Config class, to decouple Config class from the underlying
RPC implementation.

Bug: 24194427
TEST=Build apmanager for both Brillo and Chrome OS

Change-Id: I61d46c37bac875ce4a196dcc4abb752f5880ec50